Create Calculator
Created a simple calculator. The operations it  can do is below.
Addition
Subtraction
Multiplication
Division and we can delete a number and we can clear all the data when ever we suppose to do check any other results and if we gives the wrong input calculator will display as error.